# docs/swagger.yaml
basePath: /
definitions:
  resource.InputComment:
    properties:
      message:
        example: good foto!
        type: string
      photo_id:
        example: 1
        type: integer
    required:
    - message
    - photo_id
    type: object
...
  resource.UpdateComment:
    properties:
      message:
        example: amazing photo
        type: string
    required:
    - message
    type: object
...
info:
  contact:
    email: kadekbintanga@gmail.com
    name: API Support
  description: This is a service for final project
  license:
    name: Apache 2.0
  termsOfService: http://swagger.io/terms/
  title: Final Project by Bintang
  version: "1.0"
paths:
...
  /api/v1/comment:
    post:
      consumes:
      - application/json
      description: Create a photo with the input payload
      parameters:
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/resource.InputComment'
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a comment
      tags:
      - Comment
  /api/v1/comment/{comment_id}:
    delete:
      consumes:
      - application/json
      description: delete a comment with the token
      parameters:
      - description: Comment ID
        in: path
        name: comment_id
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a comment
      tags:
      - Comment
    put:
      consumes:
      - application/json
      description: Update a comment with the input payload
      parameters:
      - description: Comment ID
        in: path
        name: comment_id
        required: true
        type: string
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/resource.UpdateComment'
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update a comment
      tags:
      - Comment
  /api/v1/comment/{photo_id}:
    get:
      description: Get comment with bearer token
      parameters:
      - description: Photo ID
        in: path
        name: photo_id
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get comment
      tags:
      - Comment
...
swagger: "2.0"